// src/config/enums/configCodenames.enum.ts
import { registerEnumType } from '@nestjs/graphql';

export enum ConfigCodeNames {
  DEFAULT_SAT = 'DEFAULT_SAT',
  TIME_ZONES = 'TIME_ZONES',
  MAX_APPOINTMENTS_PER_MONTH = 'MAX_APPOINTMENTS_PER_MONTH',
  MAX_DISTANTE_COACH_AVAILABITY_QUERY = 'MAX_DISTANTE_COACH_AVAILABITY_QUERY',
  MAX_DISTANTE_COACH_APPOINTMENT = 'MAX_DISTANTE_COACH_APPOINTMENT',
  MIN_COACHING_SESSION_DURATION = 'MIN_COACHING_SESSION_DURATION', // 30 min
  MAX_COACHING_SESSION_DURATION = 'MAX_COACHING_SESSION_DURATION', // 120 min
  MAX_COACHES_SUGGESTIONS = 'MAX_COACHES_SUGGESTIONS', // 3 times
  COACH_SUGGESTED_BY_REQUEST = 'COACH_SUGGESTED_BY_REQUEST', // 3 coaches
  DEFAULT_COACHING_SESSION_FEEDBACK = 'DEFAULT_COACHING_SESSION_FEEDBACK',
  DAYS_NUMBER_ASSIGNED_COACHEE_AS_RECENT = 'DAYS_NUMBER_ASSIGNED_COACHEE_AS_RECENT', // SOLO INGRESAR NUMEROS COMO STRING EJ '3',
  DAYS_COACHEE_RECENT_REGISTERED = 'DAYS_COACHEE_RECENT_REGISTERED', // CANT DE DIAS EN LA QUE SE CONSIDERA AL COACHEE COMO REGISTRADO "RECIENTEMENTE" '3',
  DAYS_COACHEE_WITHOUT_ACTIVITY = 'DAYS_COACHEE_WITHOUT_ACTIVITY', // CANT DE DIAS EN LA Q SE CONSIDERA QUE UN COACHEE NO TIENE ACTIVIDAD RECIENTE EJ'21',
  DEFAULT_SESSION_DURATION = 'DEFAULT_SESSION_DURATION', // Para definir la cantidad de minutos que dura una sesión
}

registerEnumType(ConfigCodeNames, {
  name: 'ConfigCodeNames',
});